// frontend/src/api/generated/model/statusEventInterfaces.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */
import type { RuntimeInterfaceInventoryResponse } from './runtimeInterfaceInventoryResponse';
import type { StatusEventInterfacesType } from './statusEventInterfacesType';

export interface StatusEventInterfaces {
  type: StatusEventInterfacesType;
  data: RuntimeInterfaceInventoryResponse;
}