// frontend/src/api/generated/model/statusEventOutbounds.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */
import type { RuntimeOutboundsResponse } from './runtimeOutboundsResponse';
import type { StatusEventOutboundsType } from './statusEventOutboundsType';

export interface StatusEventOutbounds {
  type: StatusEventOutboundsType;
  data: RuntimeOutboundsResponse;
}