// frontend/src/api/generated/model/statusEventService.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */
import type { HealthResponse } from './healthResponse';
import type { StatusEventServiceType } from './statusEventServiceType';

export interface StatusEventService {
  type: StatusEventServiceType;
  data: HealthResponse;
}