// frontend/src/api/generated/model/statusEventSnapshot.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */
import type { StatusEventSnapshotData } from './statusEventSnapshotData';
import type { StatusEventSnapshotType } from './statusEventSnapshotType';

export interface StatusEventSnapshot {
  type: StatusEventSnapshotType;
  data: StatusEventSnapshotData;
}